instead of int type ids injects objects via constructor, moves handlers to adapter namespace, renames some commands and their handlers
