--- src/PrestaShopBundle/Resources/config/services/core/domain/meta.yml
@@ -1,25 +1,0 @@
-services:
-    _defaults:
-        public: true
-
-    # COMMAND HANDLERS
-
-    prestashop.core.domain.meta.command_handler.save_meta:
-        class: 'PrestaShop\PrestaShop\Core\Domain\Meta\CommandHandler\SaveMetaHandler'
-        tags:
-            - name: 'tactician.handler'
-              command: 'PrestaShop\PrestaShop\Core\Domain\Meta\Command\SaveMetaCommand'
-
-    prestashop.core.domain.meta.command_handler.edit_meta:
-        class: 'PrestaShop\PrestaShop\Core\Domain\Meta\CommandHandler\EditMetaHandler'
-        tags:
-            - name: 'tactician.handler'
-              command: 'PrestaShop\PrestaShop\Core\Domain\Meta\Command\EditMetaCommand'
-
-    # QUERY HANDLERS
-
-    prestashop.core.domain.meta.query_handler.get_meta_for_editing:
-        class: 'PrestaShop\PrestaShop\Core\Domain\Meta\QueryHandler\GetMetaForEditingHandler'
-        tags:
-            - name: 'tactician.handler'
-              command: 'PrestaShop\PrestaShop\Core\Domain\Meta\Query\GetMetaForEditing'
